% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddWeights.R
\name{AddWeights}
\alias{AddWeights}
\title{Add transformed weights into cci}
\usage{
AddWeights(cci, features = c("means_e", "FC_ec"), w = c(1, 1))
}
\arguments{
\item{cci}{The dataframe extracted from  DataTransformCPD function}

\item{features}{The selected features that can be included into the final weights, default: c('means_e','FC_ec')}

\item{w}{The weights controls features' contribution, default: c(1,1)}
}
\value{
The new cci with added weights
}
\description{
Add transformed weights into cci
}
\examples{
\donttest{
cci <- AddWeights(cci,features=c('means_e','FC_ec'),w = c(1,1))
}

}
