% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTransformCPD.R
\name{DataTransformCPD}
\alias{DataTransformCPD}
\title{merge the expr and ctrl CellPhoneDB data and get the differential cross-talking cell clusters}
\usage{
DataTransformCPD(
  exp_dir,
  ct_dir,
  celltypes = c("Vn_ACKR1", "Mph_C2_GPNMB", "PB_MZB1", "CD4_C5_FOXP3", "Mo_C2_EREG",
    "Mast_TPSB2", "FB_C2_POSTN"),
  SelfInter = FALSE,
  FC_cutoff = 1.5,
  pvalue_cutoff = 0.05,
  FC_constant = 0.01
)
}
\arguments{
\item{exp_dir}{CellPhoneDB output directory of expr group, like cellphonedb/expr/out}

\item{ct_dir}{CellPhoneDB output directory of ctrl group, like cellphonedb/ctrl/out}

\item{celltypes}{the cell types of interest}

\item{SelfInter}{whether self-self communications are included. Default: FALSE, not included}

\item{FC_constant}{small constant to avoid issues associated with division by zero, default: 0.01}

\item{FC_cutoff:}{foldchange (expr/out) cutoff for the differential communication, default:1.5}

\item{pvalue_cutoff:}{pvalue cutoff (extracted from CellPhoneDB) of expr group, default: 0.05}
}
\value{
A dataframe that includes the differential ligand-receptor interactions in expr vs ctrl
}
\description{
This function extracts the results from CellPhoneDB; merges the expr and ctrl results based
on cell-cell L-R interactions; calculates the foldchange (expr/ctrl) based on the LR expression;
and get the differential cross-talking cell clusters in the expr group compared to ctrl.
}
\examples{
\donttest{
cci <- DataTransformCPD('cellphonedb/ccm/out','cellphonedb/ep/out',celltypes=c('Tcell','Bcell'))
}

}
