% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistGraph.R
\name{DistGraph}
\alias{DistGraph}
\title{choose the different strategies to calculate the distance}
\usage{
DistGraph(graph, method = "shortest.paths", weight_name = "weight", a = 0.5)
}
\arguments{
\item{graph}{An graph created by igraph package}

\item{method}{'shortest.paths' or 'diffusion'}

\item{weight_name}{Either NULL or a character string giving an edge attribute name}

\item{a}{The parameter controls the likelihood that the diffusion process follows a connection originating from a node (default is 0.5)}
}
\value{
distance matrix of the nodes in the graph
}
\description{
choose the different strategies to calculate the distance
}
\examples{
\donttest{

}

}
