% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlots.R
\name{EqarrowPlot}
\alias{EqarrowPlot}
\title{Eqarrow Plot}
\usage{
EqarrowPlot(
  graph,
  layout,
  edge.width,
  edge.lty = rep(1, ecount(graph)),
  edge.arrow.size = rep(1, ecount(graph)),
  vertex.shape = "circle",
  edge.curved = autocurve.edges(graph),
  ...
)
}
\arguments{
\item{graph}{An graph created by igraph package}

\item{layout}{layout}

\item{edge.width}{edge.width}

\item{edge.lty}{edge.lty}

\item{edge.arrow.size}{edge.arrow.size}

\item{vertex.shape}{vertex.shape}

\item{edge.curved}{edge.curved}

\item{...}{other parameters}
}
\value{
draw Eqarrow Plot
}
\description{
Eqarrow Plot
}
