% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindComb.R
\name{FindComb}
\alias{FindComb}
\title{find the ligand-receptor(LR) combinations}
\usage{
FindComb(cci, n = 2)
}
\arguments{
\item{cci}{The dataframe extracted from  DataTransformCPD function}

\item{n}{Maximum number of LR pairs in one combination, default:2}
}
\value{
the ligand-receptor(LR) combination list
}
\description{
find the ligand-receptor(LR) combinations
}
\examples{
\donttest{
lr_combination <- FindComb(cci,n=2)
}

}
