% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDiffusionDis.R
\name{FindDiffusionDis}
\alias{FindDiffusionDis}
\title{calculate diffusion-kernel based distance}
\usage{
FindDiffusionDis(graph, weight_name = "weight", a = 0.5)
}
\arguments{
\item{graph}{An graph created by igraph package}

\item{weight_name}{Either NULL or a character string giving an edge attribute name}

\item{a}{The parameter controls the likelihood that the diffusion process follows a connection originating from a node (default is 0.5)}
}
\value{
distance matrix of the nodes in the graph
}
\description{
calculate diffusion-kernel based distance
}
\examples{
\donttest{

}

}
