% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSPDis.R
\name{FindSPDis}
\alias{FindSPDis}
\title{calculate shortest path distance based on Dijkstra's algorithm}
\usage{
FindSPDis(graph)
}
\arguments{
\item{graph}{An graph created by igraph package}
}
\value{
distance matrix of the nodes in the graph
}
\description{
calculate shortest path distance based on Dijkstra's algorithm
}
\examples{
\donttest{

}

}
