% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LollipopPlot.R
\name{LollipopPlot}
\alias{LollipopPlot}
\title{Plotting ParsimonyC3 results with Lollipop Plot}
\usage{
LollipopPlot(df, top = 10)
}
\arguments{
\item{top}{The top number of LR combinations, default: 10}

\item{The}{results from ParsimonyC3}
}
\value{
Plotting the priortized LR pair combinations
}
\description{
Plotting ParsimonyC3 results with Lollipop Plot
}
\examples{
\donttest{
LollipopPlot(result,top = 10)
}

}
