% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlots.R
\name{NetworkPlot}
\alias{NetworkPlot}
\title{Plotting ParsimonyC3 results with Network Plot}
\usage{
NetworkPlot(cci, start = "Vn_ACKR1", lr_select = "GRN_SORT1|PECAM1_CD38")
}
\arguments{
\item{cci}{The dataframe with LR interactions}

\item{start}{the starting celltype name}

\item{lr_select}{The selected LR combination}
}
\value{
Plotting the selected LR combination
}
\description{
Plotting ParsimonyC3 results with Network Plot
}
\examples{
\donttest{
NetworkPlot(cci,lr_select='PECAM1_CD38|integrin_a4b7_complex_FN1')
}

}
