% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParsimonyC3.R
\name{ParsimonyC3}
\alias{ParsimonyC3}
\title{ParsimonyC3 methods}
\usage{
ParsimonyC3(cci, lr_combination, start = "Vn_ACKR1", b = 0.02, r = 1)
}
\arguments{
\item{cci}{The dataframe with LR interactions}

\item{lr_combination}{The ligand-receptor(LR) combination list}

\item{start}{the starting celltype name}

\item{b}{A parameter that adjusts the relative strength in the shortest path mode}

\item{r}{A parameter that adjusts the relative strength in the diffusion-kernel based mode}
}
\value{
The priortized LR pair combinations. c_lr: LR combination ID; z_shortest: z score (can be found in the Paper Methods) in shortest path mode;
z_diffusion: z score in diffusion-kernel based mode; n_lr: number of LRs in the specific combination
}
\description{
ParsimonyC3 methods
}
\examples{
\donttest{
result <- ParsimonyC3(cci,lr_combination)
}

}
