% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkPlots.R
\name{ShortestpathPlot}
\alias{ShortestpathPlot}
\title{ShortestpathPlot}
\usage{
ShortestpathPlot(graph, df_sub, plot_list)
}
\arguments{
\item{graph}{An graph created by igraph package}

\item{df_sub}{The dataframe with the LR combinations}

\item{plot_list}{The paths between the starting node to the remaining nodes}
}
\value{
A Shortestpath Plot
}
\description{
ShortestpathPlot
}
