% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightTransform.R
\name{WeightTransform}
\alias{WeightTransform}
\title{weight transformation methods}
\usage{
WeightTransform(x, transform = "neg_log_normal")
}
\arguments{
\item{x}{input weight vector}

\item{transform}{transformation methods, default:'neg_log_normal': -1\emph{log10(x/max(x));
'neg_log': -1}log10(x);'exp_neg_log':exp(1)^(-1*log10(x));'log':log10(x)}
}
\value{
transformed weight vector
}
\description{
weight transformation methods
}
\examples{
\donttest{

}

}
